+++ include/MyRK5.h
@@ -16,7 +16,7 @@ class MyRK5 final : public MyODEIntegrator
 	MyRK5() = default;
 	~MyRK5() override = default;
 
-	void step(const float& t, const float& h, arma::Col<float>& x, MyFunctor& f) const override;
+	void step(const double& t, const double& h, arma::Col<double>& x, MyFunctor& f) const override;
 };
