+++ source/MyDoublePendulum.cpp
@@ -4,20 +4,20 @@
 
 #include "MyDoublePendulum.h"
 
-arma::Col<float>  MySimpleDoublePendulum::eval(const arma::Col<float> &x) const
+arma::Col<double>  MySimpleDoublePendulum::eval(const arma::Col<double> &x) const
 {
-    const float cos_d = std::cos(x[0]-x[1]);
-    const float sin_d = std::sin(x[0]-x[1]);
-    const float den = 2 - cos_d*cos_d;
-    const float den2 = den*den;
-    const float prod = x[2]*x[3];
-    const float num = sin_d*( (x[2]*x[2] + 2*x[3]*x[3])*cos_d - prod*(2+cos_d*cos_d));
-    const float f2 = num / den2;
+    const double cos_d = std::cos(x[0]-x[1]);
+    const double sin_d = std::sin(x[0]-x[1]);
+    const double den = 2 - cos_d*cos_d;
+    const double den2 = den*den;
+    const double prod = x[2]*x[3];
+    const double num = sin_d*( (x[2]*x[2] + 2*x[3]*x[3])*cos_d - prod*(2+cos_d*cos_d));
+    const double f2 = num / den2;
 
     return arma::Col {
         (-   x[2] + x[3]*cos_d) /den,
         (- 2*x[3] + x[2]*cos_d) /den,
-        static_cast<float>(- f2 + 2*M_G *std::sin(x[0])),
-        static_cast<float>(  f2 +   M_G *std::sin(x[1]))
+        - f2 + 2*M_G *std::sin(x[0]),
+          f2 +   M_G *std::sin(x[1])
     };
 }
