+++ source/MyRK8.cpp
@@ -7,11 +7,11 @@
 namespace
 {
 	// Coefficients for the Butcher tableau for RK8
-	constexpr float m_A[8] = {
+	constexpr double m_A[8] = {
 		0.125, 0.25, 0.375, 0.5, 0.625, 0.75, 0.875, 1.0
 	}; // m_A is here just for future implementations with time dependence
 
-	constexpr float m_B[8][8] = {
+	constexpr double m_B[8][8] = {
 		{0.125},
 		{0.041666666666666664, 0.20833333333333334},
 		{0.020833333333333332, -0.08333333333333333, 0.4375},
@@ -22,13 +22,13 @@ namespace
 		{0.025, 0.0, 0.2, 0.16666666666666666, -0.125, 0.075, 0.43333333333333335, 0.225}
 	};
 
-	constexpr float m_C[9] = {
+	constexpr double m_C[9] = {
 		0.041666666666666664, 0.0, 0.16666666666666666, 0.20833333333333334, -0.125, 0.25, 0.225, 0.23333333333333334, 0.0
 	};
 }
 
-void MyRK8::step(const float& t, const float& h, arma::Col<float>& x, MyFunctor& f) const {
-	arma::Col<float> k[9];
+void MyRK8::step(const double& t, const double& h, arma::Col<double>& x, MyFunctor& f) const {
+	arma::Col<double> k[9];
 
 	k[0] = h * f(x);
 	for (int i = 1; i < 9; i++) {
